fix relative fopen() of piscsi rom
the problem relates to the relative path open for the piscsi rom file,
which obviously broke when i moved the filesystem around. for some
unknown reason i could still boot from piscsi so it wasn't immediately
apparent to me that this had occurred. `make clean` to the rescue.

relative paths will be a long term issue, especially if binaries are
installed in locations such as `/usr/bin`, so it would be nice to fix
this. one way of remedying the issue that i considered was putting rom
files into `/usr/share/pistorm/*` but this seemed overkill for just one
rom at present. furthermore this would likely need defining in the
config file. this is likely undesirable as it would add an additional
config step to users.

the fix, at the moment, is to encode `piscsi.rom` as a static buffer
and change the `fopen` to `fmemopen`. `lsegout.bin` is still being
written to a relative location. the *uaes bury their roms within the
emulator binary, and whilst pistorm is not uae, the images are fairly
small. `hexdump` and a format string to the rescue.